correction dans le session back et front dans parcelle et cultures

+++ src/Controller/ParcelleController.php
@@ -16,13 +16,26 @@ use Symfony\Component\HttpClient\HttpClient;
 use Symfony\Contracts\HttpClient\Exception\ClientExceptionInterface;
 use Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface;
 use Symfony\Component\Filesystem\Exception\IOExceptionInterface;
+use App\Entity\User;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
+
 
 #[Route('/parcelle')]
 final class ParcelleController extends AbstractController
 {
     #[Route(name: 'app_parcelle_index', methods: ['GET'])]
-    public function index(Request $request, ParcelleRepository $parcelleRepository): Response
+    public function index(Request $request, ParcelleRepository $parcelleRepository, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
+        $loggedInUserId = $session->get('client_user_id');
+        
+        if (!$loggedInUserId) {
+            return $this->redirectToRoute('app_user_login');
+        }
+        $loggedInUser = $entityManager->getRepository(User::class)->find($loggedInUserId);
+        if (!$loggedInUser) {
+            return $this->redirectToRoute('app_user_login');
+        }
+
         $page = (int) $request->query->get('page', 1); // Get the current page from the query parameters
         $limit = 9; // Number of parcelles per page
         $offset = ($page - 1) * $limit;
@@ -42,13 +55,25 @@ final class ParcelleController extends AbstractController
             'currentPage' => $page,
             'totalPages' => $totalPages,
             'MAPBOX_API_KEY' => $mapboxApiKey, // Pass to Twig
+            'loggedInUser' => $loggedInUser,
+
         ]);
     }
     
 
     #[Route('/new', name: 'app_parcelle_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, SessionInterface $sessionr): Response
     {
+        $loggedInUserId = $session->get('client_user_id');
+        
+        if (!$loggedInUserId) {
+            return $this->redirectToRoute('app_user_login');
+        }
+        $loggedInUser = $entityManager->getRepository(User::class)->find($loggedInUserId);
+        if (!$loggedInUser) {
+            return $this->redirectToRoute('app_user_login');
+        }
+
         $parcelle = new Parcelle();
         $form = $this->createForm(ParcelleType::class, $parcelle);
         $form->handleRequest($request);
@@ -64,23 +89,47 @@ final class ParcelleController extends AbstractController
         return $this->render('parcelle/new.html.twig', [
             'parcelle' => $parcelle,
             'form' => $form,
+            'loggedInUser' => $loggedInUser,
+
         ]);
     }
 
     #[Route('/{id}', name: 'app_parcelle_show', methods: ['GET'])]
-    public function show(Parcelle $parcelle): Response
+    public function show(Parcelle $parcelle, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
+        $loggedInUserId = $session->get('client_user_id');
+        
+        if (!$loggedInUserId) {
+            return $this->redirectToRoute('app_user_login');
+        }
+        $loggedInUser = $entityManager->getRepository(User::class)->find($loggedInUserId);
+        if (!$loggedInUser) {
+            return $this->redirectToRoute('app_user_login');
+        }
+
         $mapboxApiKey = $_ENV['MAPBOX_API_KEY']; // Load from .env
 
         return $this->render('parcelle/show.html.twig', [
             'parcelle' => $parcelle,
             'MAPBOX_API_KEY' => $mapboxApiKey,
+            'loggedInUser' => $loggedInUser,
+
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_parcelle_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Parcelle $parcelle, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Parcelle $parcelle, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
+        $loggedInUserId = $session->get('client_user_id');
+        
+        if (!$loggedInUserId) {
+            return $this->redirectToRoute('app_user_login');
+        }
+        $loggedInUser = $entityManager->getRepository(User::class)->find($loggedInUserId);
+        if (!$loggedInUser) {
+            return $this->redirectToRoute('app_user_login');
+        }
+
         $form = $this->createForm(ParcelleType::class, $parcelle);
         $form->handleRequest($request);
 
@@ -93,12 +142,23 @@ final class ParcelleController extends AbstractController
         return $this->render('parcelle/edit.html.twig', [
             'parcelle' => $parcelle,
             'form' => $form,
+            'loggedInUser' => $loggedInUser,
+
         ]);
     }
 
     #[Route('/{id}', name: 'app_parcelle_delete', methods: ['POST'])]
-    public function delete(Request $request, Parcelle $parcelle, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Parcelle $parcelle, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
+        $loggedInUserId = $session->get('admin_user_id');
+
+        if (!$loggedInUserId) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
+        $loggedInUser = $entityManager->getRepository(User::class)->find($loggedInUserId);
+        if (!$loggedInUser) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
         if ($this->isCsrfTokenValid('delete'.$parcelle->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($parcelle);
             $entityManager->flush();
@@ -110,8 +170,18 @@ final class ParcelleController extends AbstractController
 
     //ADMIN
     #[Route('/dashboard', name: 'app_parcelle_index2', methods: ['GET'])]
-    public function dashparcelle(Request $request, ParcelleRepository $parcelleRepository): Response
+    public function dashparcelle(Request $request, ParcelleRepository $parcelleRepository, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
+        $loggedInUserId = $session->get('admin_user_id');
+
+        if (!$loggedInUserId) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
+        $loggedInUser = $entityManager->getRepository(User::class)->find($loggedInUserId);
+        if (!$loggedInUser) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
+
         $query = $request->query->get('q', '');
         $parcelles = !empty($query) ? $parcelleRepository->searchByName($query) : $parcelleRepository->findAll();
         
@@ -141,6 +211,7 @@ final class ParcelleController extends AbstractController
             'parcelles' => $parcelles,
             'MAPBOX_API_KEY' => $mapboxApiKey,
             'query' => $query,
+            'loggedInUser' => $loggedInUser,
         ]);
     }
     
@@ -151,8 +222,17 @@ final class ParcelleController extends AbstractController
     
 
     #[Route('/dashboard/new', name: 'app_parcelle_new2', methods: ['GET', 'POST'])]
-    public function new2(Request $request, EntityManagerInterface $entityManager): Response
+    public function new2(Request $request, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
+        $loggedInUserId = $session->get('admin_user_id');
+
+        if (!$loggedInUserId) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
+        $loggedInUser = $entityManager->getRepository(User::class)->find($loggedInUserId);
+        if (!$loggedInUser) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
         $parcelle = new Parcelle();
         $form = $this->createForm(ParcelleType::class, $parcelle);
         $form->handleRequest($request);
@@ -168,23 +248,46 @@ final class ParcelleController extends AbstractController
         return $this->render('parcelle/newAdmin.html.twig', [
             'parcelle' => $parcelle,
             'form' => $form,
+            'loggedInUser' => $loggedInUser,
+
         ]);
     }
 
     #[Route('/dashboard/{id}', name: 'app_parcelle_show2', methods: ['GET'])]
-    public function show2(Parcelle $parcelle): Response
+    public function show2(Parcelle $parcelle, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
+        $loggedInUserId = $session->get('admin_user_id');
+
+        if (!$loggedInUserId) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
+        $loggedInUser = $entityManager->getRepository(User::class)->find($loggedInUserId);
+        if (!$loggedInUser) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
         $mapboxApiKey = $_ENV['MAPBOX_API_KEY']; // Load from .env
 
         return $this->render('parcelle/showAdmin.html.twig', [
             'parcelle' => $parcelle,
             'MAPBOX_API_KEY' => $mapboxApiKey,
+            'loggedInUser' => $loggedInUser,
+
         ]);
     }
 
     #[Route('/dashboard/{id}/edit', name: 'app_parcelle_edit2', methods: ['GET', 'POST'])]
-    public function edit2(Request $request, Parcelle $parcelle, EntityManagerInterface $entityManager): Response
+    public function edit2(Request $request, Parcelle $parcelle, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
+        $loggedInUserId = $session->get('admin_user_id');
+
+        if (!$loggedInUserId) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
+        $loggedInUser = $entityManager->getRepository(User::class)->find($loggedInUserId);
+        if (!$loggedInUser) {
+            return $this->redirectToRoute('app_user_loginback');
+        }
+
         $form = $this->createForm(ParcelleType::class, $parcelle);
         $form->handleRequest($request);
 
@@ -197,11 +300,13 @@ final class ParcelleController extends AbstractController
         return $this->render('parcelle/editAdmin.html.twig', [
             'parcelle' => $parcelle,
             'form' => $form,
+            'loggedInUser' => $loggedInUser,
+
         ]);
     }
 
     #[Route('/dashboard/{id}', name: 'app_parcelle_delete2', methods: ['POST'])]
-    public function delete2(Request $request, Parcelle $parcelle, EntityManagerInterface $entityManager): Response
+    public function delete2(Request $request, Parcelle $parcelle, EntityManagerInterface $entityManager, SessionInterface $session): Response
     {
         if ($this->isCsrfTokenValid('delete'.$parcelle->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($parcelle);
